+++ Code/C#/VoltageFeedbackMessages/GetVoltageThresholdResponse.cs
@@ -9,11 +9,11 @@ namespace VoltageFeedbackMessages.Requests
     [DataMemberNamesClass(typeof(GetVoltageThresholdResponseDataMemberNames))]
     public class GetVoltageThresholdResponse
     {
-        private double _Voltage;
+        private float _Voltage;
         [JsonPropertyName(GetVoltageThresholdResponseDataMemberNames.Voltage)]
         [JsonInclude]
         [DataMember(Name = GetVoltageThresholdResponseDataMemberNames.Voltage)]
-        public double Voltage { get { return _Voltage; } protected set { _Voltage = value; } }
+        public float Voltage { get { return _Voltage; } protected set { _Voltage = value; } }
     }
 }
