+++ Code/C#/VoltageFeedbackMessages/SetVoltageThresholdRequest.cs
@@ -8,11 +8,11 @@ namespace VoltageFeedbackMessages.Requests
     [DataMemberNamesClass(typeof(SetVoltageThresholdRequestDataMemberNames))]
     public class SetVoltageThresholdRequest
     {
-        private double _Voltage;
+        private float _Voltage;
         [JsonPropertyName(SetVoltageThresholdRequestDataMemberNames.Voltage)]
         [JsonInclude]
         [DataMember(Name = SetVoltageThresholdRequestDataMemberNames.Voltage)]
-        public double Voltage { get { return _Voltage; } protected set { _Voltage = value; } }
+        public float Voltage { get { return _Voltage; } protected set { _Voltage = value; } }
     }
 }
