+++ Code/C#/VoltageFeedbackMessages/VoltageMessage.cs
@@ -9,11 +9,11 @@ namespace VoltageFeedbackMessages.Messages
     [DataMemberNamesClass(typeof(VoltageMessageDataMemberNames))]
     public class VoltageMessage
     {
-        private double _Voltage;
+        private float _Voltage;
         [JsonPropertyName(VoltageMessageDataMemberNames.Voltage)]
         [JsonInclude]
         [DataMember(Name = VoltageMessageDataMemberNames.Voltage)]
-        public double Voltage { get { return _Voltage; } protected set { _Voltage = value; } }
+        public float Voltage { get { return _Voltage; } protected set { _Voltage = value; } }
     }
 }
