+++ Code/C++/HVPSController/main/IO/SoftStartHandler.cpp
@@ -8,18 +8,18 @@
 #include "Timing/Delay.hpp"
 #include "Timing/TimeHelper.hpp"
 
-double SoftStartHandler::doSoftStart(const HVPSConfiguration& config1, const HVPSConfiguration& config2){
+float SoftStartHandler::doSoftStart(const HVPSConfiguration& config1, const HVPSConfiguration& config2){
 	Outputs::setSoftStartResistorBypassOnOff(false);
-	double voltages[WINDOW_SAMPLES];
+	float voltages[WINDOW_SAMPLES];
 	size_t index = 0;
-	double lastAverageVoltage=0.0;
+	float lastAverageVoltage=0.0f;
 	uint64_t startTimeMs = TimeHelper::ms();
 	uint64_t talkNext = startTimeMs +TALK_INTERVAL_MILLISECONDS;
 	bool reachedMinimumVoltage = false;
 	bool voltageStoppedIncreasing = false;
-	double result = -1;
+	float result = -1.0f;
 	Inputs::useADCPowerSupplyVoltageFeedbackChannel([&](IADCSession&& adc){
-		double minimumVoltageCanRead = ADC::getMinimumVoltageCanRead();
+		float minimumVoltageCanRead = ADC::getMinimumVoltageCanRead();
 		Log::Info(TAG, "Minimum voltage can read was: %f", minimumVoltageCanRead);
 		while(true){
 			Delay::ms(SAMPLE_INTERVAL_MS);
@@ -32,11 +32,11 @@ double SoftStartHandler::doSoftStart(const HVPSConfiguration& config1, const HVP
 				continue;
 			}
 			index=0;
-			double averageVoltage = ArrayHelper::average(voltages, WINDOW_SAMPLES);
+			float averageVoltage = ArrayHelper::average(voltages, WINDOW_SAMPLES);
 			Log::Info(TAG, "averageVoltage %f", averageVoltage);
-			double dVoltage = averageVoltage - lastAverageVoltage;
+			float dVoltage = averageVoltage - lastAverageVoltage;
 			reachedMinimumVoltage = averageVoltage>=MINIMUM_VOLTAGE_TO_REACH;
-			voltageStoppedIncreasing = dVoltage <=0.0;
+			voltageStoppedIncreasing = dVoltage <=0.0f;
 			if(reachedMinimumVoltage&&voltageStoppedIncreasing){
 				Log::Info(TAG, "Reached minimum voltage and voltage stopped increasing", lastAverageVoltage);
 				Outputs::setSoftStartResistorBypassOnOff(true);
